:recycle: Small Refactor to help add close buttons to tabs
