+++ src/main/kotlin/com/github/sloppylopez/moneypennyideaplugin/actions/SendToPromptFileFolderTreeAction.kt
@@ -6,7 +6,6 @@ import com.intellij.openapi.actionSystem.*
 import com.intellij.openapi.components.Service
 import com.intellij.openapi.components.service
 import com.intellij.openapi.diagnostic.thisLogger
-import com.intellij.openapi.fileEditor.FileEditorManager
 import com.intellij.openapi.project.Project
 import javax.swing.ImageIcon
