+++ src/main/kotlin/com/github/sloppylopez/moneypennyideaplugin/toolWindow/TextAreaFactory.kt
@@ -1,3 +1,4 @@
+package com.github.sloppylopez.moneypennyideaplugin.toolWindow
 import com.github.sloppylopez.moneypennyideaplugin.components.MoneyPennyTextArea
 import com.intellij.notification.*
 import com.intellij.openapi.components.Service
